QMetaObject::invokeMethod is called incorrectly in DataEngine::connectSource / DataEngine::connectAllSources
also: If a DataEngine fails to create the source in
sourceRequested(source), an empty source is created in d->source(source)
and then used during QMetaObject::invokeMethod, which calls updated for
the applet.
Better: don't create the source. This forces DataEngines to create the
source.

Patch by Michael Olbrich <michael-olbrich@web.de>

svn path=/trunk/KDE/kdebase/workspace/libs/plasma/; revision=695097