Fix the calendar dialog popping up in the wrong place the first time.
The issue is that we are using a QProxyWidget, and it seems that QWidgets don't have a sensible size() until they're shown for the first time.  So we use sizeHint() (or effectiveSizeHint() in this case) instead.

The only issue I can see is that if the widget passed to the dialog is in a layout, its size() and effectiveSizeHint(Qt::PreferredSize) might not match.  Is this something we need to deal with?  If so, how?

CCMAIL: plasma-devel@kde.org


svn path=/trunk/KDE/kdebase/workspace/libs/plasma/; revision=859660