Fix Plasma ScrollBar flickableItem position binding when mouse or keys are used
The RangeModel value was being incremented directly by the mouse and keys
events, and it was overriding the binding with the flickableItem position.
Once the ScrollBar button was clicked, the direct manipulation of the
flickable component was not updating the scroll position.

The new function 'incrementValue' now does the correct work.

Signed-off-by: Daker Fernandes Pinheiro <dakerfp@gmail.com>