fix right side task item shadow gradient
this is another bugfix in the form of "how about we stop showing
horrible vertical lines on the taskbar entries!". left side still needs
done. what makes this difficult is they're extremely small RASTER files.
so i have to zoom in 9000% and can only select pixels that are huge. not
the best to do when you are working with gradients.