Plasma QML components: fix internal anchoring in Button and ToolButton
There is a bug in the tool button and button components resulting in
layout breakage if one clears and sets the text property of the
component when not visible, see the attached screenshot.

I have tried to solve the issue without changing the existing
anchoring system, but without success. The only working solution
was to put the icon and the label item into Row item. That way I was
able to fix the bug and even get rid of the ugly explicit
non-declarative anchor assignments.

I have also removed the preferredWidth property of the label item,
that one always evaluated to paintedWidth, anyway.

REVIEW: 107813