make GenericPackage the superclass for Plasmoid
the hierarchy was a bit inverted; this allows GenericPackages to
carry a few less Plasmoid-specific items without changing any other
behaviour.

uses of Plasma/Generic will need to bear in mind that they are not
using Plasmoids. this opens doors to more appropriate and easier
flexibility for other package types we ough to have, e.g. Wallpaper