expose Applet::id() in the plasmoid object
We need this in the containment to do our book-keeping, if we just guess
the applet id, we don't know which containment is passed into
addApplet(), so we can't know where to restore it. So we use applet.id
to match a geometry saved to the config with the applet that is being
loaded.