Add Plasma/Generic servicetype
This irons out an asymmetry: Plasma/Generic was never read from
KService, as the service type wasn't registered, we just looked it up in
the right directory.

This patch allows us to use KService to list packages, the
installPackage cmake macro installs the service files for us already.
Apps that look up the package in the "right path" still work.