Add the "theme" directory to the plasmoid directory structure
This adds an extra step to the lookup of SVG images created with the
Svg() global function: after looking in images/ in the plasmoid and then
in the desktop theme, it looks in theme/$DESKTOP_THEME_NAME and then
theme/ in the plasmoid.

This allows plasmoid authors to add images that they want to allow theme
authors to override.

Forward-port of review request 109857.