Revert "Enabling C++11 flags for clang and gcc"
This reverts commit 7c5e2e49ae5bef1d4071ff1ec555bd7b7beb61c5.

In KF5, C++11 is not to be assumed on all our platforms. So don't force
it from cmake. That said we have to make sure our code builds in both
case (C++11 available or not). I'll get in touch with the build.kde.org
admins to sort that out.

CCMAIL: ivan.cukic@kde.org