Fix empty group operations
There appears to be a subtle change in KConfig*, which means that if a
config group is empty, it won't be listed in groupsMap. Since operations
files can have empty groups, which are simply calls without arguments,
these would not be listed, and thus not be callable at all.

Fix: Add a dummy item per group, not beautiful, but effectively makes it
work.