Remove windowFlags property
This is a needless duplication of the property, and it causes bugs.

Removing it, and porting all users to the flags property from QWindow
magically makes layering and resizability bugs vanish.

One problem is left: the dialog dismisses too easily, it should not
dismiss when being resized, moved or when a child dialog or QMenu opens
from it.