[plasma-shell] Add an option to suppress any output (--shut-up)
The amount of warnings that plasma-shell has, makes it super hard to
make use of tools like journalctl or to grep ~/.xsession-errors.

We need these tools to diagnose possible bugs in the session start or
any other software that redirects stderr or those places.

We can remove this option once all the Warnings are fixed, specially
the one in Qt: https://codereview.qt-project.org/#change,73943

CCMAIL: plasma-devel@kde.org