We shouldn't take "lock widgets" from corona, but only from containment
It is up to the Containment to decide if the user is allowed or not
to lock/unluck the widgets, so corona should not add one when there is none
(user is not allow) and it shouldn't add another one when there is already
one.

Fixes duplicated action in toolbox.