Add test for PlasmaCore.ToolTip.target property
This uses a MouseEventListener and a MouseArea to test whether the
tooltip redirects events, and tests a common implementation scenario
with one shared tooltip.

This test (willingly) exposes a problem that with the MouseArea, we
filter out all mouse events, and the target loses input. I.e. the last
image in the row doesn't receive the click.