Add OSD notifications to plasma-shell
As we decided to make the OSD available only for certain things, the API
is intentionally limited to very particular set of things. The list is
not complete yet I believe, but it's definitely a finite and a small set

Also as it's not a public service, I left it under org.kde.plasma_shell
on D-Bus.

And we're still missing all the icons except volume, but Jens is on it.

REVIEW: 115301