Set minimumWidth of button to actually show the text
The minimumWidth of a plasma button is based on the paintedWidth
of the text item. There is no way to work out the correct width it
should be if we set the text to elide.
(except for the hack of using a hidden fake text item)

We should always fit contents. This won't cause any problems unless a
developer has a broken layout, at which point that should be fixed

REVIEW: 115500