Return intersection of KWS::workArea() and QScreen::availableGeometry()
KWindowSystem::workArea() gives us the whole workarea, which in case of dual screens is the whole virtual desktop. QApplication::desktop()->availableGeometry(id) gives the full screen geometry for the given screen. So we just intersect them and have a proper availableScreenRect().

This is just a temporary method until the patch in [1] is merged in Qt
and released, then we can revert back to the commented method.

[1] - https://codereview.qt-project.org/#change,80606

BUG: 331968