Disable elide in TabButton
This messes up the fontmetrics needed to size the tab button. Before,
the labels would be elided no matter what.

This is more a band-aid than a real fix, we'll need to investigate how
we can make the sizing work correctly, but at least it "unscrews" the
layout of all of our tabbars.