Revert "Setting parent on toolbox causes problems. Instead delete via connection"
This reverts commit c38e6a204e66a2fca2f9f0606475731b5c6bf0d1.

This causes problems on the panel toolbox.
We were already trying to set a parent; however as it's a property of a QQuickItem
it only changes the visual parent; kdeclarative has a code path that if there's
a visual parent to skip the memory management; bug is probably there instead.