Don't leak
we were creating new packagestructures on each package creation
and never deleting them, causing a pretty big leak.

now keep a single instance of the structures, with prototype packages.
members are static for binary compatibility.
i would go for max one bic change in this library that's when and if
this library will be made pubic.

Patch by Aaron Seigo