Use font metrics to scale icons for high dpi outputs
QScreen, through EDID reports bogus values for physicalDotsPerInch*().
This leads to oversized icons on monitors with bogus edid information.

This patch changes the ratio underlying to the icon sizing for displays
with a DPI > 96 * 1.5 to use the ratio between the font as rendered (its
effective pixelSize, essentially) to scale the icon sizes up.

As we rely on proper font metrics throughout already, this should bring
sizing in line with the font, which is something that makes sense as it
means we're sharing the underlying mechanism (font metrics) for sizing
in different areas.

The downside of this patch is that we're essentially working around an
issue that should be fixed in the hardware, the monitor's edid.
Unrealistic.