Only search in relevant subdirectories
As packages are installed in subdirectories based on their servicetype,
it makes sense to only consider those directories. This speeds up
searching considerably, since it doesn't have to be recursive, meaning
we won't find in dirs more than one level deeper as the current
directory.

- add a hashmap with subdirectories per servicetype, so we know where to
  look in the most common cases
- add a separate function for non-recursive queries. The recursive one
  can be removed at some point, right now it's useful to gauge the
  performance impact

The subdirectories should actually come from the servicetype, as
hard-coding them in the pluginloader means we can only query for known
servicetypes. (It might still make sense to cache at least the most
common ones here, so we can skip reading in service files in many cases,
just to find the right subdirectory.)