Make-it-work
This program runs over a subdirectory of $SHARE/plasma and creates a
cache file of all packages contained in that directory. The cache file
is saved in Qt's binary json format for efficient reading and writing.
This tool should be run whenever a package is installed into the
respective subdirectory.

The cache is currently used in the plugintrader autotest.

In detail:

- binary renamed to plasma-update-packagecache
- Read metadata.desktop file and convert it to json
- return q QVariantMap from convert, and append it to our json data
- output file defaults to plasma-update-packagecache.json in the
  packages subdirectory

Missing:
- translations in json file
- smarter handling of string list keys (currently hardcoded to known
  ones)
- optimizations all over, this code "works", but is very brittle, still
- locking, while traversing a subdirectory, we want to lock it to
  prevent other processes, which might run simultaneously, to overwrite
  our file mid-flight