Removed the call of clearDataForMimeJob on a null pointer
tjob pointer is checked for being null (result of the dynamic_cast)
and then passed to clearDataForMimeJob. The issue can be solved by
replacing tjob with job in the problematic line, but since the
mimeTypeRetrieved method called later already processes jobs that
are not KIO::TransferJob, there is no purpose in doing it here as well.

Change-Id: I863ac41ad66ac3cbceca82adeab85498315a5e42