Fix the platformstatus kded module
The module .desktop file has a typo in the X-KDE-DBus-ModuleName which
results in the KCM not being able to see the status of the module.

Plus it renames the .desktop file to not contain the kded_ prefix, now
it matches every other kded .desktop file we ship (and also matches the
library and dbus module names).

Change-Id: I5006672ca041d9820b4388ba4e99b7899f36be18