Fix fallback from invalid L&F packages
The mechanism should have deleted the old key from the config if it was
invalid and then try again in order to get the default. As we were
deleting the wrong key (actually the value of the shell package key),
this was turning into a recursive loop. Deleting the correct key
obviously fixes this.

Thanks Bille for reporting!

Reviewed-by: Marco Martin

BUG:343510