try to reposition before showing
to minimize dialog repositions, try to put it at the proper position before showing.

besides making things a bit snappier,
showing at 0,0 for an instant has some interesting sideeffects:
MouseArea has a rather strange way to manage hoverevents:
at window show, it will check if the item is under the absolute mouse position,
if so, will emit an entered() signal. This causes the Kickoff menu to
automatically select the last tab on first show, because for an instant
(that never gets drawn on the screen) all the tabs will actually be under the mouse cursor

Change-Id: Ie9fc45f2a5076c3f49b42e8a7cee109b04db46b9