Unify when tooltips are shown and kept alive.
A new tooltip would only be shown if it had content to show.

However in the old code if the tooltip was already visible we would keep
the tooltip alive and sync with empty contents regardless of whether we
had anything to show.

This seems to clear up the system tray where we have nested tooltip
areas for the tray applet and the icon itself and things would flicker

Change-Id: I7d9ded9edb70672c572e008c8d6f1ffe2d76b773
Changelog: Avoid showing an empty tooltip