make PlasmaCore.ColorScope work better as advertised
Since attached properties cannot ever change during the lifetime of the object,
a colorscope needs to be created for each obect that asks for the
attached property. now each instance if (internal) has m_inherit
(only those created as an attached property), will search for the property
of the proper ancestor. Not super efficient, but the proeprties should be
actually asked for rarely enough.

doesn't look much nice, but seems to work pretty reliably and the only way to.

Change-Id: I735da5551dbb8c2641be88d245f998d00ed977ac
REVIEW:122988